Fetch official site insecurely
curl is otherwise failing on the Github setup:

---
curl: (60) SSL certificate problem: unable to get local issuer certificate
More details here: https://curl.haxx.se/docs/sslcerts.html

curl failed to verify the legitimacy of the server and therefore could not
establish a secure connection to it. To learn more about this situation and
how to fix it, please visit the web page mentioned above.
Error: Process completed with exit code 60.
---

This isn't the ideal fix, but should be a Good Enough workaround as
we're not doing anything with the results other than manually comparing
them to Wikidata.